fix : 배포 삭제 수정, test : 배포 상태 변경 테스트

+++ deployment-service/src/main/java/com/ttalkak/deployment/deployment/application/inputport/DeleteDeploymentInputPort.java
@@ -1,8 +1,10 @@
 package com.ttalkak.deployment.deployment.application.inputport;
 
 import com.ttalkak.deployment.deployment.application.outputport.DeploymentOutputPort;
+import com.ttalkak.deployment.deployment.application.outputport.ProjectOutputPort;
 import com.ttalkak.deployment.deployment.application.usecase.DeleteDeploymentUsecase;
 import com.ttalkak.deployment.deployment.domain.model.DeploymentEntity;
+import com.ttalkak.deployment.deployment.framework.projectadapter.dto.ProjectInfoResponse;
 import com.ttalkak.deployment.deployment.framework.web.request.DeploymentDeleteRequest;
 import com.ttalkak.deployment.common.global.error.ErrorCode;
 import com.ttalkak.deployment.common.global.exception.EntityNotFoundException;
@@ -20,10 +22,16 @@ public class DeleteDeploymentInputPort implements DeleteDeploymentUsecase {
 
     private final DeploymentOutputPort deploymentOutputPort;
 
+    private final ProjectOutputPort projectOutputPort;
     @Override
-    public void deleteDeployment(Long deploymentId) {
+    public void deleteDeployment(Long userId, Long deploymentId) {
         DeploymentEntity deploymentEntity = deploymentOutputPort.findDeployment(deploymentId)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorCode.NOT_EXISTS_DEPLOYMENT));
+        Long projectId = deploymentEntity.getProjectId();
+        ProjectInfoResponse projectInfo = projectOutputPort.getProjectInfo(projectId);
+        if(userId != projectInfo.getUserId()){
+            throw new EntityNotFoundException(ErrorCode.NOT_EXISTS_DEPLOYMENT);
+        }
         deploymentEntity.deleteDeployment();
         deploymentOutputPort.save(deploymentEntity);
     }
